INTYG-7429: Lagt till intygSentEvent i LogJob

+++ service/src/main/java/se/inera/statistics/service/processlog/intygsent/IntygsentLogConsumer.java
@@ -20,6 +20,6 @@ package se.inera.statistics.service.processlog.intygsent;
 
 public interface IntygsentLogConsumer {
 
-    long processBatch();
+    int processBatch();
 
 }
